Simplify builders by marking default variants for enums

+++ src/models/madhab.rs
@@ -9,16 +9,18 @@
 /// For Hanafi madhab, Asr is roughly an hour later than the Shafi madhab.
 use serde::{Deserialize, Serialize};
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
+#[derive(Debug, Default, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Deserialize, Serialize)]
+#[repr(u8)]
 pub enum Madhab {
+    #[default]
     Shafi = 1,
     Hanafi = 2,
 }
 
 impl Madhab {
     #[must_use]
-    pub const fn shadow(self) -> i32 {
-        self as i32
+    pub const fn shadow(self) -> u8 {
+        self as u8
     }
 }
 
@@ -31,7 +33,7 @@ mod tests {
     #[rstest]
     #[case::from_shafi(Madhab::Shafi, 1)]
     #[case::from_hanafi(Madhab::Hanafi, 2)]
-    fn test_shadow_length(#[case] madhab: Madhab, #[case] length: i32) {
+    fn test_shadow_length(#[case] madhab: Madhab, #[case] length: u8) {
         assert_eq!(madhab.shadow(), length);
     }
 }
